added Runtime tests, refactoring & cleanup
