+++ Assets/UnitTests/Editor/EditorTests.cs
@@ -1,0 +1,4 @@
+/// <summary>
+/// Fortunately we can simply use the runtime tests class as base for the editor tests.
+/// </summary>
+public class EditorTests : RuntimeTests {}
