--- Assets/UnitTests/Runtime/RuntimeEventStressTests.cs
@@ -1,13 +1,0 @@
-using NUnit.Framework;
-using TestHelper;
-using Unity.Entities;
-using Unity.PerformanceTesting;
-using Unity.Transforms;
-
-public partial class RuntimeEventStressTests : ECSTestsFixture
-{
-    [Test, Performance]
-    public void EntitiesTestExample()
-    {
-    }
-}
