--- Assets/UnitTests/Utility/TestHelper.asmdef
@@ -1,25 +1,0 @@
-{
-    "name": "TestHelper",
-    "references": [
-        "Unity.Entities",
-        "Unity.Collections",
-        "Unity.Jobs",
-        "Unity.Burst",
-        "Unity.Mathematics",
-        "Unity.Transforms",
-        "Unity.Collections.BurstCompatibilityGen"
-    ],
-    "optionalUnityReferences": [
-       "TestAssemblies"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": true,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "noEngineReferences": false
-}
